+++ src/eca/llm_providers/google.clj
@@ -1,0 +1,31 @@
+(ns eca.llm-providers.google
+  (:require
+   [clojure.string :as string]
+   [eca.config :as config]
+   [eca.features.login :as f.login]
+   [eca.shared :refer [multi-str]]))
+
+(defmethod f.login/login-step ["google" :login/start] [{:keys [db* chat-id provider send-msg!]}]
+  (swap! db* assoc-in [:chats chat-id :login-provider] provider)
+  (swap! db* assoc-in [:auth provider] {:step :login/waiting-login-method})
+  (send-msg! (multi-str "Now, inform the login method:"
+                        ""
+                        "manual: Manually enter API Key")))
+
+(defmethod f.login/login-step ["google" :login/waiting-login-method] [{:keys [db* input provider send-msg!]}]
+  (case input
+    "manual"
+    (do
+      (swap! db* assoc-in [:auth provider] {:step :login/waiting-api-key
+                                            :mode :manual})
+      (send-msg! "Paste your google API Key"))
+    (send-msg! (format "Unknown login method '%s'. Inform one of the options: max, console, manual" input))))
+
+(defmethod f.login/login-step ["google" :login/waiting-api-key] [{:keys [db* input provider send-msg!] :as ctx}]
+  (if (not (string/blank? input))
+    (do
+      (config/update-global-config! {:providers {"google" {:key input}}})
+      (swap! db* update :auth dissoc provider)
+      (send-msg! (format "API key saved to %s" (.getCanonicalPath (config/global-config-file))))
+      (f.login/login-done! ctx))
+    (send-msg! (format "Invalid API key '%s'" input))))